requirements.txt
base58==2.1.1
cffi==1.15.1
pycparser==2.21
PyNaCl==1.5.0
